requirements.txt
# 核心框架
fastapi==0.135.3
uvicorn==0.44.0
gradio==6.12.0

# LangChain 相关
langchain==1.2.15
langchain-community==0.4.1
langchain-text-splitters==1.1.2

# 向量存储和嵌入
chromadb==1.5.8
FlagEmbedding==1.3.5

# 文档处理
pypdf==6.10.2

# API 和工具
openai==2.32.0
python-dotenv==1.2.2
python-multipart==0.0.26
requests==2.33.1

# 深度学习（FlagEmbedding 依赖）
torch==2.11.0
transformers==4.36.0
huggingface_hub==1.10.2